+++ R/params_to_observations.R
@@ -69,9 +69,7 @@ params_to_observations <- function(params) {
             mutate(
               longitude = known$known_lon[id],
               latitude = known$known_lat[id],
-              observation_comments = glue::glue(
-                "Automatically computed from `known$stap_id==1` with `params_to_observations()`"
-              )
+              observation_comments = "Automatically computed from `known$stap_id==1` with `params_to_observations()`"
             )
         }
 
@@ -81,9 +79,7 @@ params_to_observations <- function(params) {
             mutate(
               longitude = known$known_lon[id],
               latitude = known$known_lat[id],
-              observation_comments = glue::glue(
-                "Automatically computed from `known$stap_id==-1` with `params_to_observations()`"
-              )
+              observation_comments = "Automatically computed from `known$stap_id==-1` with `params_to_observations()`"
             )
         }
       }